ulxly: rename agglayer to aggkit; remove commented code;
