--- cmd/ulxly/bridge_service/agglayer/l1_info_tree.go
@@ -1,13 +1,0 @@
-package agglayer
-
-// type getInjectedL1InfoLeafResponse struct {
-// 	BlockNum          uint64 `json:"block_num"`
-// 	BlockPos          uint64 `json:"block_pos"`
-// 	GlobalExitRoot    string `json:"global_exit_root"`
-// 	Hash              string `json:"hash"`
-// 	L1InfoTreeIndex   uint64 `json:"l1_info_tree_index"`
-// 	MainnetExitRoot   string `json:"mainnet_exit_root"`
-// 	PreviousBlockHash string `json:"previous_block_hash"`
-// 	RollupExitRoot    string `json:"rollup_exit_root"`
-// 	Timestamp         uint64 `json:"timestamp"`
-// }
