+++ cmd/ulxly/bridge_service/evidence/native_legacy_bridge_manual_test.md
@@ -1,4 +1,4 @@
-# Agglayer Bridge Service Manual Test
+# Aggkit Bridge Service Manual Test
 
 ## Create account
