+++ cmd/ulxly/bridge_service/factory/factory.go
@@ -2,7 +2,7 @@ package bridge_service_factory
 
 import (
 	"github.com/0xPolygon/polygon-cli/cmd/ulxly/bridge_service"
-	"github.com/0xPolygon/polygon-cli/cmd/ulxly/bridge_service/agglayer"
+	"github.com/0xPolygon/polygon-cli/cmd/ulxly/bridge_service/aggkit"
 	"github.com/0xPolygon/polygon-cli/cmd/ulxly/bridge_service/legacy"
 )
 
@@ -10,5 +10,5 @@ func NewBridgeService(url string, insecure, useLegacy bool) (bridge_service.Brid
 	if useLegacy {
 		return legacy.NewBridgeService(url, insecure)
 	}
-	return agglayer.NewBridgeService(url, insecure)
+	return aggkit.NewBridgeService(url, insecure)
 }
